requirements.txt
fastapi
uvicorn
gunicorn
torch
torchvision
timm
pillow
opencv-python-headless
numpy
huggingface_hub
scikit-learn